Decouple the firmware pool from the device manager
The reason is that the device manager in usbprog-core doesn't necessary
have the firmwarepool from usbprog available. So we have to decouple the
two libraries completely.

We add a new class UpdateDevice. Each firmware has a member of that type.
We can give a vector of UpdateDevice to the device manager when detecting
update devices.


Signed-off-by: Bernhard Walle <bernhard@bwalle.de>